finagle: Change threshold AC filter EMA window size parameter from Long to Duration
Problems

1. The threshold AC filter's EMA window size parameter is a unitless `Long`.
This is difficult to reason about.
2. When the filter processes a request it increases its EMA timestamp by 1.
Therefore the timestamp counts the number of generated requests. The problem
with this is that the EMA is updated to the same final value when it receives a
burst of 1000 requests over 1 ms as when it receives 1000 requets over 1 minute
(for example). In other words, the EMA's value is not sensitive to the time
delta between each request. Respecting time deltas between requests is important
in reality, because we want recent requests to influence the EMA more than older
requests.

Solutions

1. Update the window size parameter to be a `Duration` instead of a `Long`.
2. Use a timestamp based on the actual time instead of on the number of
generated requests.

RB_ID=862392