finagle/finagle-core: Propagate retries on a request in the Context
Problem

We'd like to be able to record the number of retries on a request
from the server's perspective. In order to do this, the number
of retries need to be propagated.

Solution

Propagate the number of retries (from one level up) in the Context
of a request.

RB_ID=862640