finagle-http: fix connection management for non-keepalive requests.
Problem

When issuing HTTP/1.0 or non-keep-alive requests (i.e. with Connection: close),
the connection may be marked as closable before a request/response has
completed.  Specifically, when finagle-http's ConnectionManager observes a
non-keepalive request, it may satisfy onClose before a response is observed.
This causes the entire connection to be torn down (and the response canceled)
before the response can be received.

Solution

Consider a connection as closable only when an equal number of requests and
responses have been observed.

Fixes twitter/finagle#549