finagle-http: Make StreamingTest generic and provide a netty4 implementation
Problem

We have nice end to end streaming tests, but they're only for netty3
for now.

Solution

Make them abstract so that they can be used with other implementations.

Result

We now have a solid streaming test suite for netty4 http, which
helps us be more confident in our netty4 http implementation.  Once
we add support for using http2 without the handshake works, we'll
be able to use the abstract test suite for that too.

RB_ID=867121