finagle/finagle-thriftmux: Remove gauges in ThriftEmulator
Problem

We create gauges per-connection in ThriftEmulator
on the hotpath. Gauge creation can be slow if gauge cleanup
is happening on another thread, which stalls the connection.

Solution

Remove the gauges in ThriftEmulator, because it already contains
counters for the same stats. If, later, we decide that these gauges
are useful, they will need to be moved to the ListeningServer and
references passed into ThriftEmulator. This way, the gauges are tied
to the lifecycle of the ListeningServer, and not created per-connection
in ThriftEmulator.

RB_ID=867459