finagle-netty4: Fix I/O thread leak
Problem

rb/865672 introduced a bug by scaling the number of I/O workers
with clients and servers. Each `NioEventLoopGroup` will create
a fixed amount of threads and we create one instance per client
and server.

Solution

Change the `WorkerPool` param to create the `EventLoopGroup`. This
leaks netty4 types, but it's in the finagle-netty4 package so it
seems legitimate.

RB_ID=867862