finagle-http: Expand HttpNackFilter to also handle non-retryable nacks
Problem

All Nacks are treated as Restartable from finagle-http's point of view, but it
is sometimes desirable to signal that a Nack should not be retried.

Solution

Expand HttpNackFilter to handle non-retryable Nacks

Result

The "finagle-http-nonretryable-nack" header is used to transmit this information
over HTTP. Clients and servers now support & handle this header.

RB_ID=865468