finagle-stats: Add toggle for histogram export
Problem

By default, finagle-stats exports full histogram details even when
there were no data points recorded. We want to change this behavior,
and we want to do it in a controlled manner.

Solution

Introduce a toggle, "com.twitter.finagle.stats.exportEmptyHistograms"
which we will use to change the default behavior in the future.

RB_ID=868317