finagle-http2: Implement client-side streaming
Problem

We want to be able to reuse the http/1.1 netty pipeline, but some
of the netty objects don't have a place to put the stream id, so
we can only use the subset of netty http objects that have headers.
We also need each stream to be able to use the http/1.1 netty
pipeline separately, so that aggregation and disaggregation works
properly–imagine a handler accidentally aggregating over several
different pipelines.

Solution

We can get around this by using a wrapper object that has space for
a stream id, and then unwrapping the wrapper object, sending it
through a synthetic netty pipeline just for that stream, and then
rewrapping the wrapper object.  By having a synthetitc netty pipeline
per stream, this also solves our problem around aggregation and
disaggregation.

While we were in here, we moved the SchemifyingHandler to only be
added on upgrade, which simplifies the downgrade path.  I also
figured out that a test that seemed wrong was actually right, but
I had screwed up aggregation, which was why it was breaking.

Result

Client-side streaming works, which means that http/2 now works end
to end in finagle.

RB_ID=866174