finagle-http: Introduce toggle for treating 500s as failures
Problem

`finagle-http` uses the default `ResponseClassifier` and as such does
not treat server error status codes (500-599) as failures. This in
turn means without user effort, the default behavior for stats is
surprising and the client behavior for failure accrual is suboptimal.

Solution

Introduce a toggle, "com.twitter.finagle.http.serverErrorsAsFailures",
so that this functionality can be changed in a controlled manner.

RB_ID=869303