finagle-http2: Removed byte handlers from http2 frame pipeline
Problem

After constructing the child handlers in the http/2 server-side
codec, we then add all of the same initializers to each of the child
handlers.  This works OK, until we turn on SSL, at which point SSL
dies because it expects bytes and doesn't know what to do with
http/2 frames.  We should make sure that none of the handlers which
expect bytes are in the http/2 child handlers.

Solution

Separate the byte handlers and the domain object handlers into two
separate ChannelInitializers, and only pass the domain object ones
to the user.

Result

The child handlers have all of the stuff we expect them to have,
and none of the stuff we don't expect them to have!

RB_ID=868869