finagle-mux: reassemble tcp segments in wireshark mux dissector
Problem

The mux wireshark plugin fails to classify fragmented mux messages.
This limits the practical utility of the dissector as all messages
in excess of 1500 bytes are fragmented by the MTU imposed by ethernet.

Solution

Use the dissect_tcp_pdus api to desegment the fragmented mux messages.
This patch also add the "mux overhead" field to the dissector which
trivializes graphing the network efficiency of mux traffic.

RB_ID=871474