finagle-http2: Allow connections to start with prior knowledge
Problem

We have many tests for http that don't convert well to http/2
cleartext, because in http/2 cleartext the first request is secretly
an http/1.1 request.  This means that instead of testing the http/2
implementation, we're still secretly testing the http/1.1 implementation.

Solution

The http/2 spec describes an alternate way of starting an http/2
connection, called "prior knowledge".  Both client and server must
know that they are speaking prior-knowledge http/2, and then they
can skip the upgrade step.

We introduce a PriorKnowledge configuration option, and when we
turn it on we remove the upgrade step from our pipelines.  On the
client-side, we can also simplify our transporter, which takes much
of its inherent complexity from having to judge whether each
connection is upgraded, so we've introduced the PriorKnowledgeTransporter,
a thin wrapper around an unbounded FutureCache[SocketAddress,
MultiplexedTransporter].

We've also added a ChannelHandler that
buffers requests to the prior knowledge http/2 client-side pipeline,
to avoid a race between when clients call `write` and the handler
has `channelActive` called on it, which is when it writes the
connection preface.

Result

We can reuse http tests like AbstractStreamingTest in our http/2
suite, and in the future, we will also be able to interoperate with
other http implementations that only support prior knowledge, like
grpc in particular.

RB_ID=870052