finagle-http: Revert "Fix ResponseProxy for chunked transfers"
Problem

The change in RB 851799 actually causes the stated problem in
classes that extend ResponseProxy (as opposed to instantiating
an anonymous ResponseProxy in a filter) and thus breaks the
`c.t.inject.server.ResponseBuilder#EnrichedResponse` for chunked
responses in Finatra.

Solution

Revert the change and add a test that shows how to make the
intended use in a filter work along with a new test that tests
subclasses of ResponseProxy which fails without the revert
to ResponseProxy.

RB_ID=871611