finagle-mux: Wrap BadMessageException in Failure
Problem

The message from BadMessageException doesn't offer any details about the
peer that sent the message, making it more difficult to debug.

Solution

Wrap BadMessageException in Failure so that information about the peer
can be set by the filter chain and will be manifest in the exception
message.

RB_ID=870309