finagle-http: HttpNackFilter shouldnt return a body on HEAD request
Problem

The c.t.f.h.f.HttpNackFilter doesn't consider the request method when
forming a response from a Failure, which is a problem for HEAD
requests because the HttpNackFilter includes a body generated response,
which is a protocol error for a response to a HEAD request.

Solution

Don't add a body if the nacked request was a HEAD request.

RB_ID=872242