finagle-thriftmux: Allow ThriftMux.Clients to be filtered.
Problem

ThriftMux.Clients do not currently support a `filtered` method.
This is supported in the the `StdStackClient` but ThriftMux.Client
is a `StackBasedClient` (a superclass of the `StdStackClient`).

Solution

Implement a `filtered` method directly on ThriftMux.Client.

Result

Users will be able to directly filter on the ThriftMux.Client
in the same manner supported by Thrift.Client.

RB_ID=874560