finagle-thrift: abstract transport implementation
Problem

The finagle-thrift project is structured in a way that
exposes many netty3 types and has a package layout that makes
it difficult to add a netty4 implementation.

Solution

Restructure the package to remove netty3 types from most of
the public interface and move the netty3 specifics to a new
package, com.twitter.finagle.thrift.transport.netty3. The
notable exceptions are the ThriftServerFramedCodec and
ThriftClientFramedCodec which are both part of the public API
and widely used at this time.
The types in the c.t.f.thrift.legacy package have been removed
or moved to the netty3 specific package. The only public types
that were removed were those in Types.scala.

RB_ID=873982