finagle-netty4: Introduce receive_buffer_bytes (only available with pooling)
Problem

While pooling of receive buffers is still experimental, it's important to provide
a means for users to tune that properly. One of the most important inputs for the
tuning is the distribution of the sizes of the receive buffers.

Solution

Add a new stat `receive_buffer_bytes` (only available when pooling is enabled) that
keeps track of a histogram of the receive buffer sizes.

RB_ID=877080