Problem
com.twitter.finagle.http.filter.StatsFilter Is missing a corresponding module
so it cannot be conveniently added to a Stack.

Solution

Added the module, added the withHttpStats methods to Client and Server.

Result

It is now possible to effortlessly enable Http specific statistics on both
http servers and clients.

 Signed-off-by: Jillian Crossley <jillianc@twitter.com>

RB_ID=878465