source: Upgrade jackson to 2.8.3
Problem

We're beginning to start thinking about 2.12.x, now that 2.12.0-RC1 is a thing.
jackson-module-scala has been published for 2.12.0-RC1, but not in the version
that we're on today.

Solution

Let's move source to 2.8.3!

Result

We can pull in an ostrich PR that starts cross-publishing for scala 2.12.0-RC1.

RB_ID=878518