finagle-core: Add a NonRetryable flag to c.t.f.Failure
Problem

There is currently no mechanism to signal that a failure should not be retried.
The existing c.t.f.Failure's Restartable flag is masked off to users and its
purpose is only to signal that Finagle is confident a request can be requeued.

Solution

Add an explicit non-retryable flag which is respected by Retry/Requeue filters.
Additionally, the HttpNackFilter, HttpClientDispatcher, and
NackAdmissionControlFilter are modified to respect and use the new flag. Future
work will propagate this flag across protocol boundaries.

RB_ID=878766