finagle-http: Ensure Connection: close header is present when the connection is closed
Problem

The ConnectionManager is responsible for ensuring the persistence of an
HTTP/1.x connection, partially mediated by the Connection header.
However, when the ConnectionManager decides that it needs to close the
connection it doesn't ensure the 'Connection: close' header is set. This
may end up in an unexpected hang-up for the clients perspective.

Solution

Ensure the 'Connection: close' header is set if the ConnectionManager
intends to close the connection.

RB_ID=880280