finagle-netty4: Change pooled allocator chunk size from 2mb to 128kb
Problem / Solution

Our default value for a chunk size of Netty's pooled allocator
is unrealistically big (2mb), which results in increased memory
footprint for applications with pooling enabled.

After running a couple of experiments and identifying the maximum
receive buffer we might want to allocate in production, we decided
to change chunk size to 128kb, which should be enough for many
applications.

Note that even if the new value doesn't cover some extreme cases,
the worst thing that could happen (when a requested size is bigger
than a chunk size) is falling back to unpooled allocator (which we
do today anyway).

RB_ID=881639