finagle-core: IgnorantTrustManager now extends X509ExtendedTrustManager
Problem

Finagle's IgnorantTrustManager, which is used with the
'withTlsWithoutValidation' API for testing purposes, fails to properly
bypass certificate validation in some cases in Java 8.

Solution

Change IgnorantTrustManager to extend X509ExtendedTrustManager instead
of X509TrustManager. Implementing the methods that are part of the
X509ExtendedTrustManager interface fixes the issue for Finagle users.

RB_ID=881660