finagle-http: Use Keep-Alive to close connection on bad request
Problem

The HttpServerDispatcher closes the connection when responding to a bad
request by calling `this.close()` before returning the response. This
causes two problems:

- The semantics of `this.close()` should close the dispatcher and this the
transport immediately but due to an implementation detail that doesn't
happen and the response is allowed to pass through. This blocks fixing the
dispatcher behavior.
- The peer doesn't get any warning that the connection is going to close.

Solution

Using the keep-alive semantics to close the connection. Instead of directly
call `.this.close()` when we receive a bad request we attach a
'Connection: close' header to the generated response which will cause the
ConnectionManager to close the connection.

RB_ID=880007