finagle-netty4: Drain writes when proxy handshake is done
Problem / Solution

There is a bug in our implementation of `Netty4ProxyConnectHandler` where
we simply drops pending writes (if there are any) instead of draining them
when the handshake is done.

Note, it's quite unlikely to hit that given that Finagle never writes into
a channel that's not yet connected. Although, some of the Netty handlers
may do that (eg: SSL/TLS handler).

RB_ID=882438