finagle-http: Gracefully handle ill formed responses to a HEAD request
Problem

Finagle doesn't handle HEAD requests as it should. Specifically, the
response to a HEAD request should never have a body, but we don't enforce
that.

Solution

Make HttpServerDispatcher aware of HEAD request semantics by preparing
the response for the netty pipeline. This is done in a mandatory Filter
which has subsumed most of the conformance logic that was previously in the
`handle` method of the dispatcher. The semantics that must be adhered to
downstream are:

* The encoder MUST NOT add Content-Length or Content-Encoding headers
* A Message that is not flagged as 'chunked' is the final element in
the HTTP response and the encoder MUST be in a state that will properly
interpret the next element as the next response.

Result

Responses to HEAD requests should no longer result in a HTTP protocol
error.

RB_ID=872106