finagle-benchmark: Refactor load balancer simulation to allow for multiple clients
Problem

It's useful to be able to run our local load balancer simulation
with more than one client. For example, we may want to test clients
operating with various load balancer configs.

Solution

Introduce the flag `nClients` to represent how many clients will be
talking to the same group of servers. This also contains a larger
cleanup of the simulation since it was becoming a bit unwieldy to
manage.

RB_ID=884044