finagle-netty4: Add Workaround for Disabled AutoRead Socks5 Connection
Problem

If AutoRead (ChannelOption.AUTO_READ) is turned off for a channel, as it
is by default for Netty4 via the Netty4Transporter.Backpressure param,
as of Netty 4.1.4, the read of the Socks5 proxy server is not kicked
off and so must be done manually.

Solution

Manually kick off the read, if the initial write is successful and the
channel configuration is set to not use AutoRead. A subsequent pull
request is being submitted to Netty (Netty Issue 5933).

RB_ID=884344