finagle-thrift: lower log level in Netty4Transport
Problem

The server transport initialization order is such that custom
initialization is done before the general initialization. The netty4
transport implementation in finagle-thrift attempts to remove the
DirectToHeapInboundHandler in the custom initialization phase when it
has not yet been added to the pipeline, which naturally fails and logs
a message at level INFO which is a little scary to those not familiar
with what it means.

Solution

Removing this handler is not strictly necessary, so the logging level
on that message should be set to DEBUG.

RB_ID=884242