finagle-http: Fix close of GenSerialServerDispatcher
Problem

The finagle-http GenSerialServerDispatcher has a broken `close`
method: if a dispatch is in progress and `close` is called the underlying
transport will not close until the dispatch completes, which may be much
later than the specified deadline (and may never happen at all) potentially
resulting in resource leaks.

Solution

Use a timer to ensure that the transport is closed within the deadline
regardless of the behavior of the Future returned by `dispatch`.

Note: additional cleanup was performed to optimize the implementation of
the dispatch loop.

RB_ID=882898