Fix error syncing
I did this because currently, if there's an error during an extended
query message flow, we must issue a `Sync` in order to get Postgres to
respond to any further messages. However, the state machine currently
does not allow for a Sync message in any state but `Connected`. This
results in a connection hang in some situations.

This change allows `Sync` during any of the extended query states.