finagle-http2 - Allow initial Http/2 settings to be threaded through to client
Problem:
It is currently not possible to specify the initial Http/2 settings used by the
client. It is also not possible to specify the "ignore max header list size"
parameter. This makes it hard to test servers as the client cannot be made to
"misbehave"

Solution:
Add configuration parameters for individually setting the various http/2
settings Al on the client. Also make it possible to set the "ignore max header
list size" on the client.

Result:
It is now possible to specify the initial http/2 settings for a client, as well
as to instruct it to ignore the max header size list requested by the server.
Currently it is not possible to specify the initial settings for servers. This
will be added in a separate review if and when some needed changes to netty
land.

RB_ID=885094