finagle-http: Swallow the body of a nack response
Problem

Http nack responses contain a body. This body is typically fully
buffered on the response, but it need not be if the client allows chunked
responses. That means that if the entire body isn't decoded in one pass,
the Response will be followed by a byte chunk and the `isChunked` flag will
be set to true. The problem is that the HttpClientDispatcher
will observe the nack response and ignore any potential body that _may_ be
coming as chunks. This results in a ClassCastException for the
subsequent request because it finds a body chunk instead of a Response
in the pipeline.

Solution

Swallow the body.

RB_ID=884081