finagle-http2 - Allow client to specify stream dependency and weight
Problem:
Netty's HttpToHttp2ConnectionHandler allows the client to set headers for stream
dependency and weight which is then applied to the http/2 stream created.
Finagle's RichHttpToHttp2ConnectionHandler does not preserve this behavior.

Solution:
Since the writeHeaders method in HttpToHttp2ConnectionHandler which is
responsible for this logic is private, the bits of code for reading and applying
these values have been copied into RichHttpToHttp2ConnectionHandler

Result:
Clients can now specify stream dependency and weight by using the appropriate
headers.  Note that in Netty, in addition to these two headers, the stream id
itself can also be chosen via a header, but because of the way finagle uses the
stream ids, adding support for this is a much larger task

RB_ID=886324