finagle-redis: Restructure the Command API and encoding
Problem

Finagle Redis didn't get enough of attention in the past and its code base
(as one of the oldest in Finagle) became quite unmaintainable. In addition
to that it's written and structured in the style that requires some
modernization.

Solution

Let's refactor the API around `Command` to make it easy to understand and
maintain. As a side bonus, let's remove duplicated and dead code.

What was changed:

  - `RedisMessage` (a common parent for both `Command` and `Reply`) is
    removed since it adds nothing but extra complexity.
  - The encoding machinery was changed to eliminate duplicates hence
    reduce the code size we need to maintain.
  - Lot's of no-longer-relevant parts were removed.

Result

The `Command` API now consist of two methods `name` and `body` that
return `Buf`-representation of its name and body correspondingly.
This allows to reuse parts of the encoding logic in lots of places.

Around 500 lines of old Scala code were removed. Nearly zero impact
on the end-user.

RB_ID=885811