finagle-http2: Get streaming working pretty dang well.
Problem

Streaming is still sort of crummy in http/2

Solution

Add the generic StreamingTest so we can be confident that it works for http/2.
Also fix a ton of generic streaming bugs.

One notable change here is that we have a read loop now that reads as fast as it
can, so that we can detect RSTs and similar.

RB_ID=885299