finagle: Extract base HTTP pieces into finagle-base-http
Problem / Solution

While `finagle-http` implementation based on Netty 4 is nearly complete,
there is no way to smoothly switch existing clients/servers to a new transport
given that it's defined in a separate project/traget.

This change introduces a new finagle module `finagle-base-http` that provides
a common ground for both implementations (N3 and N4), which `finagle-http` now
depends on. This allows us to toggle between different transports w/o affecting
the caller site (users code).

Note: This RB is our first step towards distinct transport implementations for
HTTP. The essential goal is to make N3 and N4 modules as thin as possible and
make `finagle-base-http` Netty-free. This RB is not aiming to achieve that yet,
but makes a decent progress in that direction while minimizing the diff as much
as possible.

RB_ID=884614