finagle-http2: Disable flaky test
Problem

The test 'return 413s for chunked requests which stream too much data'
is flaky due to a race condition where the server closes the connection
without notice and the client may attempt to reuse it.

Solution

Disable the test until the fix for the netty library is released.

RB_ID=888396