twitter-server: Add ability to specify admin interface UI grouping, alias, and path for admin handlers
Problem

Some UI elements in the TwitterServer admin page are named after endpoints and
others have a more user-friendly name. In addition, it is currently not possible
for users to classify which category their admin handlers fall into.

Solution

Allow users to specify a group for their admin handler in addition to just the
pattern, as well as a user-friendly alias, and the path if necessary by adding
the Route and RouteIndex case classes.

RB_ID=886829