finagle-http2: Disable flaky test pending netty fix
Problem

There is another flaky test due to the same netty bug of hanging up on
the client without giving notice, resulting in a race condition where
the client may try to reuse the connection before detecting the close
event.

Solution

Disable the test pending the release of netty 4.1 which contains a fix.

RB_ID=889509