finagle-core: Add BufReader.remainingUntil
Problem / Solution

Decoding of text-based, character-delimited protocols (eg: Redis, Memcached)
is dramatically simplified when an API for locating the firs occurrence of the
given byte is present. (both N3 and N4 provide similar APIs).

This commit extends `BufReader` with a new method `remainingUntil` that does
this very trick and simplifies the decoding of CLRF-separated lines.

RB_ID=889344