finagle: Futures breakdown for developers
Problem

We have good docs for end-users, but none for people who might want to
contribute to finagle.  This is the first in a series of attempts to make this
better when onboarding to the finagle project.

The Futures implementation is difficult to read because it has a big surface
area and it's optimized to death.  We've historically explained it via oral
tradition, but it's time to do better.

Solution

Explain the history of Twitter futures, how they're different from the scala and
java Future implementations, and why that works well for us.

RB_ID=889272