finagle-netty4: Revert netty4 write behavior to mimic netty3
Problem

The Netty4 ChannelTransport deviated in the behavior of our netty3
implementation by examining the channels write buffer and dropping
writes if it found the to be 'unwritable'. This change in behavior was
not accounted for in the protocol implementations so it may cause subtle
problems.

Solution

Revert to the netty3 behavior while we consider mechanisms for write
channel backpressure.

RB_ID=890646