finagle-core: MethodBuilder, version 1
Problem

Users want the ability to customize aspects of a client on a per
method, or call, basis without having to create separate Finagle
clients for each.

Solution

Introduce `c.t.f.client.MethodBuilder` as an API for doing so.

Result

While this commit is only a partial step towards the final goals, this
allows clients to have dynamic total timeouts and application level
retry policies.

While the current version is experimental and `private[finagle]`, my
hope is that it will progress quickly out of this state.

RB_ID=889738