finagle-core: Remove old deadline histograms
Problem

We have histograms lying around that we used in making Deadline originally, but
that we don't need for understanding whether or not we should turn Deadline on
again in the future.

Solution

Remove them, or make them easier to parse and take advantage of the histogram
optimization.

Result

TimeoutFilter no longer needs a StatsReceiver!  We're exporting fewer metrics!

RB_ID=891380