finagle-http: Login once per Kerberos expiration in SpnegoAuthenticator
Problem

The current SpnegoAuthenticator ClientFilter/ServerFilter code creates a new
LoginContext and logs in for each request it processes. The reason the
ClientFilter/ServerFilter needs to log in is because it needs to get a TGT from
the KDC in order to authenticate to the TGS.
However, it does not need to log in every request as the TGT can persist for a
duration specified by the KDC administrator. This means there are redundant
calls going to the KDC for each request it filters.

Additionally, the ClientFilter does not run 'initSecContext' using the logged in
Subject's privileges. This means that the session ticket it receives in the
initial call will not be stored. This also results in redundant calls going to
the KDC for each request it filters.

Solution

In the load() method in 'trait JAAS', only create a context and call login()
after initialization and after expiration periods.

In the init() method in 'JAASClientSource', initContextAction using the current
Subject's privileges so the session ticket can be stored in its private
credentials.

Result

For each ServerFilter filter, there will be 1 fewer request going to the KDC
(except on initializaion and TGT expiration). For each ClientFilter filter,
there will be 2 fewer requests going to the KDC (except on initialization and
TGT expiration).

RB_ID=888228