finagle-*: Use MurmurHash3 for Toggle inputs
Problem

We have been using the ServerInfo().id.hashCode for the input to the
Toggle which determines if the server will be pseudo-randomly chosen to
toggle on behavior. Unfortunately the String.hashCode function is tuned
for speed and not entropy, yielding sequential hash codes for sequential
string representations, which is often the case in practice.

Solution

Switch to the Murmur3 hashing function for the same ServerInfo().id
String to get better distribution of hash codes even for sequential
String inputs.

RB_ID=892851