finagle-http: Fix race condition in SpnegoAuthenticator
Problem / Solution

c.t.f.http.SpnegoAuthenticator has a race condition when doing privileged
actions using a LoginContext's subject. When a TGT expires, it is possible for
a subject that is being used to be logged out by another thread. This will
destroy the credentials in the subject, causing a 'Failed to find any Kerberos
tgt' exception when it goes to do its privileged action. We expanded the lock
to cover the action done by a subject so the subject cannot be modified mid-
action.

RB_ID=893951