finagle-core: Zero-copy BufReader.readAll
Problem / Solution

Since `Buf`s are immutable no need for making a slice (a copy of the
`Buf` object) in `BufReader.readAll`.

Note that most of the `Buf` implementations will figure out that
`slice(0, length)` is a special case and return `this`. Although, this
is not a consistent behaviour.

RB_ID=894479