finagle-http2: handle prior knowledge simultaneously with 1.1 upgrade
Problem:
The current implementation of prior knowledge requires that you configure the
server to use prior knowledge exclusively.

Solution:
Add a handler before the Http codec that intercepts the first bytes off the wire
and compares with the fixed HTTP/2 client preface. If it matches a preface,
replace Http codec with an HTTP/2 codec, remove self and send a new client
preface down the pipeline. If the initial bytes do not match the preface, remove
self and send any consumed bytes down the pipeline.

Result:
We can now have a single cleartex server handle both Http 1.x, Http 1.1 to
HTTP/2 upgrades and prior knowledge HTTP/2 at the same time. This aligns better
with the HTTP/2 specs as I understand it.

RB_ID=874695