finagle-core: Introduce New Configuration Types for SSL/TLS
Problem

SSL/TLS code within Finagle for both Netty3 and Netty4 colocates
how an SSL/TLS engine is created with how it is configured.

Solution

Introduce new types which are to be used only for the configuration of
the SSL/TLS engine and are indifferent to how the engine is created.

RB_ID=890512