finagle-mysql: remove reference to head of stream in StdCursorResult
Problem

The recommended usage for the cursor API in finagle-mysql
results in OOM for large streams because the internal objects
in finagle-mysql hold on to the head of the AsyncStream.

For example:

```
client.cursor("...")
 .apply(..., ...)(...)
 // the transformed promise here stores a reference to
 // the `StdCursorResult` which keeps a reference to the
 // the head of the stream as a lazy val.
 .flatMap(cursoredResult => cursoredResult.stream)
 .foreach { asyncStream => /* operate on elements of stream */ }
```

Solution

Make the `stream` call a def so it isn't a member field of a
`StdCursorResult` instance.

RB_ID=893499