finagle-core: Improve logging levels used in DefaultMonitor
Problem

The logging level used in `c.t.f.util.DefaultMonitor` defaulted to
`FATAL` and for `HasLogLevel` exceptions it did not respect the level
if it was `INFO` or above. Both of these decisions were surprising to
end-users.

Solution

Always respect the level used in `HasLogLevel` and default to `WARNING`.

RB_ID=895983