finagle-netty4: do not cancel writes on interrupt
Problem

While migrating services to netty4, we have noticed that some services
run into strange issues (increased GC and failed health checks) when
cancelling writes. We suspect that the symptoms are a secondary effect
of the cancellation, but we haven't been able to find any probable cause.

Solution

We would like to divorce the rollout of netty 4 from write cancellations so
we revert the netty 4 `ChannelTransport` to have the same semantics as
netty 3 (i.e. no write cancellation).

RB_ID=896757