finagle: Migrate Flag Constants out of Failure to FailureFlags trait
Problem

All types of exceptions/failures should be flaggable in the way the Failure is
now. We wish to create a trait which can be mixed in to any exception. To do
this, Flags must be moved out of Failure.

Solution

Extract c.t.f.Failure's flags in to c.t.f.FailureFlags and add a FlaggedException
trait w/ a public API.

RB_ID=889371