finagle/finagle-core: More documentation in PendingRequestFilter and rejections counter
Problem

It's confusing for users when they have a pending request limit
configured, and the "/pending" stat does not suggest the client
is respecting this limit. It's also unclear what the difference
is between "/pending" and "/dispatcher/..../pending"

Solution

Add docs to clarify what stats will show the effecs of PendingRequestFilter,
and add a counter for rejections above the limit.

RB_ID=898184