finagle-redis: Decoding with Bufs
Problem / Solution

Decoding in `finagle-redis` should be Netty-agnostic and implemented
in terms of Twitter's `Buf`s.

Results

This is big PR that restructures most of the `finagle-redis` internals
as well as introduces some new concepts. As a bonus, it removes around
1000 lines of code that was never used.

Here is the summary of the changes.

**Naggati is dead**
Naggati codec is gone. The new `StageDecode` is inspired by Naggati
so it reuses a lot from the previous implementation, but it's less
powerful (not so generic) and simple enough to fit in less than a
hundred lines of code.

**Async decoding**
The decoding is now done asynchronously (see `NextStep.Accumualte`)
in the meaning is that if some parts of the array reply are missing
the decoder will still be able to decode (and buffer) it partially.
Previously, it has be returning "need more data" until the full response
is arrived.

**Further steps**
Decoding and framing are still represent a single step in the Netty
pipeline (although they are encoded in a separate, reusable concept).
For the sake of simplicity as well as reducing the diff size this RB
promotes a very simple (and less risky) approach and keeps everything
in the pipeline (as it was before). The next step is to figure if it
makes sense to move that out into `Transport` as well as bring the
Netty 4 implementation. The intention was to make this decision/discussion
(where the decoding should be done) explicit and prepare the ground for
any possible outcome.

RB_ID=894009